added today's project on html page
made a section with HTML comments <!-- for today's project --> and thought that this was a good idea.
made a link for Weeklies project, this could be researched, but it works as a standalone project!
added codepoem
this is not an atomic commit
first CSS from the visual design part of freeCodeCamp, very useful 🦔: